fixed debug MSVC warnings

+++ contrib/CppSQLite3/src/CppSQLite3.cpp
@@ -1038,11 +1038,11 @@ void CppSQLite3Statement::bind( int nParam, const double dValue )
 void CppSQLite3Statement::bind( int nParam, const SQL_BLOB& rBlob )
 {
     checkVM( );
-    size_t nRes = sqlite3_bind_blob( mpVM, nParam, (const void*)rBlob.toBlob( ), rBlob.blobSize( ), SQLITE_TRANSIENT );
+    size_t nRes = sqlite3_bind_blob( mpVM, nParam, (const void*)rBlob.toBlob( ), (int)rBlob.blobSize( ), SQLITE_TRANSIENT );
 
     if( nRes != SQLITE_OK )
     {
-        throw CppSQLite3Exception( nRes, "Error binding blob param", DONT_DELETE_MSG );
+        throw CppSQLite3Exception( (int)nRes, "Error binding blob param", DONT_DELETE_MSG );
     }
 }
